pyproject.toml
[project]
name = "cognix"
version = "0.1.0"
description = "Modular Discord bot platform with secure web dashboard"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "CogniX" }]

dependencies = [
    # Bot
    "discord.py>=2.4.0",
    "wavelink>=3.4.1",
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    # Database
    "SQLAlchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    # Config / validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    # Security
    "bcrypt>=4.1.3",
    "PyJWT[crypto]>=2.8.0",
    "pyotp>=2.9.0",
    "cryptography>=42.0.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.2.0",
    # Cache / IPC
    "redis>=5.0.4",
    # Utilities
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "psutil>=5.9.8",
    "qrcode[pil]>=7.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pip-audit>=2.7.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ASYNC", "S", "C4", "SIM", "RUF"]
ignore = ["S101", "S104"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S105", "S106"]
"database/migrations/**" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["database/migrations/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
